ORAC: ensure incorrect recipe names show error in x gui
Fixing a fault reported by Jan, where the xgui will just
hang forever with no indication of what is wrong if you typo
a  recipe name.

This fix doesn't automatically exit the gui, it just writes an
error message (in fact two messages) into the 'Error' section.
The last error message raised by the actual FatalError->throw
command is repeated when the gui is closed, which is the same
behaviour that occurs now for errors occuring during the
recipe processing.