_CALCULATE_STANDARD_: get sensible values when 1,1 is blank.
When running REDUCE_STANDARD, if makecube did not put any data
into position 1,1, the calculated integrated and peak flux
would be the bad value. This change uses _FIND_TRUE_PEAK_ to
attempt to find a sensible position to analyse instead of
assuming that 1,1 will always be useful.

For Grids and Jiggles, _FIND_TRUE_PEAK_ should use 1,1 unless
it is blank, and if it is blank use the position of the peak pixel.