SCUBA2_MAPSTATS: show the central RA and Dec in the log file.
This is the RA and Dec from JCMT::MapArea which is used to crop
the map to its requested area before starting. Note that it
shows the RA and Dec even if the input map is in Galactic
coordinates.

Also fixed the documentation of _WRITE_MAPSTATS_LOGFILE_ to be
correct.