SCUBA2: only change mode for log if exactly "scan"
In commit 2b280cb4f52659d73747bde3708c273f70bb2366 the mode written
to log files was changed to "dasiy" or "pong" if it contained "scan".
However this meant that "scan_pointing" became indistinguishable
from "scan" observations, such as flux calibrations, since both
were rewritten as "daisy".  This commit changes the behavior so that
the replacement only occurs when the mode is just "scan".