lib: Add extra recpars filtering examples
This adds a couple of multiple-header recipe parameter filtering
examples and a note about the ordering with respect to the object
specifier.

Note that the examples make use of a newly added generic header
"SCAN_PATTERN" and so will not work with the current CPAN version
of Astro::FITS::HdrTrans (1.57).