SCUBA2: allow calunits recpar to be "pW"
This is for cases where the user does not want an FCF calibration to
be applied.  Requested by KMAL.

Note: another possible way of implementing this would have been for
_REDUCE_SCAN_STEER_ to set the "NOCALIB" uhdr when given a CALUNITS
recipe parameter of "pW".  However it looks like other primitives
reset the "NOCALIB" uhdr so it seemed that allowing another option
for the "RECPAR_CALUNITS" uhdr would be more reliable.  (The "NOCALIB"
approach would also make it harder to tag the map in pW as "calib",
if that were ever necessary.)