heterodyne: Add dampening to the flat-field iterations.
Small variations converge well, but large deviations with nominal
ratios <0.8 or >1.2 can runaway as the measured fluxes will be under-
or over-estimated respectively.  Dampen the corrections by applying
just the square root of the apparent ratio in fluxes.  This we hope
will move in the corect direction without overshoot. The iterations
should take care of the convergence.

Since this change might lead to slower convergence in the good cases,
the default maximum number of iterations is increased by one. In
practice good flats converge after three iterations so the previous
limit of five might still be sufficient.  However, I have not
performed sufficient tests to know if six will be adequate.