lib: Restore full list of SCUBA-2 "da codes"
This makes the SCUBA-2 frame class's "_dacodes" private method
return ("a", "b", "c", "d") when it is not discovering the codes
from $ORAC_REMOTE_TASK.  This in turn makes "flag_from_bits" return
a list of OK files for all 4 sub-arrays unless it finds a meta
file.

I'm not sure why the full list of sub-arrays was previously
commented out.  Commits 7259699a58 and e1f8ce2a8f don't explain.
Perhaps it was an attempt at fixing the problem addressed in
commit 44dec77335 for S2SRO?