lib: Directly probe flag files for flag/skip
This combines the "orac_loop_flag" code which checks for the next
observation with that which tries to skip over missing observations.
Now the set of flag files being looked for is generated dynamically
for each candidate future observation number, which avoids the problem
of meta files needing to be re-read for SCUBA-2 (as in commit
44dec77335) in case a default set of subarrays had been assumed.

The reason for this change is that the pipeline failed to skip over
MCE errors for SCUBA-2 where some of the flag files for an observation
are written but others are not.  In that case "orac_check_data_dir"
returned the number of the partial observation, but this never
passed the "$next != $i" test to be identified as a skip.  I'm not
sure why the previous implementation called "orac_check_data_dir"
10 times in a loop -- that would involve reading the whole directory
10 times.  The new version probes directly for the expected flag files,
up to 10 observations ahead.  However it is potentially susceptible
to the problem, mentioned in a comment in the previous version, of
multiple observations appearing on disk simultaneously, but, since
the introduction of the loop mentioned above, subsequent to that
comment being added, the code would only have protected against this
problem in the first of the 10 iterations, so the new version isn't
much worse in this regard.  The new version also has the advantage
of being much simpler and not retaining a list of "next" file names.
If we ever need to start the pipeline from a SCUBA-2 observation which
hasn't been written yet and not all sub-arrays are present we may
also need to stop retaining the list of current file names and
dynamically generate it too each time around the loop so that meta
files are re-read.