SCUBA2: combine FTS-2 phase correction primitives
This combines the two phase correction primitives, which were nearly
identical, into one primitive with a mode option to select between
single- and double-sided modes.  Perhaps it should determine which
mode to use automatically, but for now the mode is specified as a
primitive argument to maintain the existing behavior of the
REDUCE_FTS_SCAN recipe (always double-sided).