SCUBA2: allow numbered suffices in FTS-2 pipeline
Modified the SCUBA-2 frame class "inout" method so that it can
(optionally) retain the numerical portion of the file name suffix.
This is to allow for the updated file naming scheme used by the
FTS2SPLIT SMURF task.  It has been made optional (controlled via
a uhdr entry) so that it only applies to the FTS-2 pipeline.

The REDUCE_FTS_ZPD recipe now generates a ZPD file with a numbered
suffix since the new mechanism is not disabled in that recipe
after _SPLIT_FTS_SCANS_ enables it.  However in the case of
scan direction-specific ZPD measurements, the numbered suffix
is probably necessary anyway.  Therefore it is probably simplest
to keep it always.  The alternative would be to remove the
number from the suffix (perhaps at the _STACK_IMAGES_BY_SUBARRAY_
step), but only in the non scan direction-specific case.