Calib: Create a new flat calibration for ACSIS.
It's been a long time since I added a calibration and SUN/233 doesn't
explain how to do this and stuff like the role of the standard
xxxxindex and xxxxnoupdate methods.  Thus this flat calibration may
not be fully correct or optimal, but it seems to work.

For now it can only be used by having an $ORAC_DATA_OUT/index.flat
until the $ORAC_DATA_CAL/index.flat is populated from the logged flats
derived during the legacy processing.

One area of doubt is whether to demand that the index file contain H00
through H15 columns, or permit just receptors listed in header
RECPTORS.  For the latter a dynamically created
$ORAC_DATA_OUT/rules.flat is needed, listing only the required
receptors.  While I expect the calibration to complain of mismatches
btween the rules and index, the new flat method does default missing
ratios to 1.0 just n case something odd slips through.

The full list is needed for the $ORAC_DATA_CAL master index.

As yet there is no validation of the ratios (say betwen 0.6 and 1.5).