heterodyne: Create a primitive to apply flat-field from a calibration index.
This is my first attempt to use the index.flat calibration file for recipe
parameter FLAT_METHOD = index.

Since flat fielding occurs in the group-processing phase where the
current Frame will always be the last one, its UT data would be used.
If data from multiple nights are members of the group, the wrong
ORACTIME would be used to select the same flat for all nights.  To
circumvent this there is a fudge with the headers to make the
calibration system pick up the desired ORACTIME for each night.  This
is modified using the UT date as decimal time is not needed.  If later
we find flats changing during a night, this will need revision.