heterodyne: Create a rules.flat file where needed.
Rejigged things so that an $ORAC_DATA_OUT/rules.flat is made if
needed. such as when there's an $ORAC_DATA_OUT/index.flat with missing
receptor columns.  There is still some reliance on the user to have a
index.flat containing all the required columns, or a matching pair if
both the index and rules files are prsent in $ORAC_DATA_OUT.

The best advice for users is make an index file with all 16 receptors
present, setting any unknown entries to 1.0.