SCUBA2: use catalog positions in _SHIFT_IMAGES_
If the observation is from a CAL / EC project, check whether we
have catalog coordinates for the given source.  Then compare these
to the WCS SkyRef position and if the difference is >= 0.1", use
the catalog coordinates instead.  However if the difference is > 60",
the SkyRef position is retained on the assumption that the wrong
object name was probably used for the observation.  Warnings are
printed to indicate when the shift reference coordinates are being
changed.