heterodyne: Revise the key for TIMESERIES_MASKS uhdr.
To avoid having to update the masks if they are to be used more than
once, revise the key to exclude the suffix.  The key therefore has the
form a<utdate>_<obsnum>_<subscan>_<subfile>, which should not change
after additional processing steps.  This permits easier repeated use
of the timeseries masks.