heterodyne: First versions of primitives to remove reference-spectrum emission.
These still need more testing with a variety of data, especially
_FIND_EMISSION_LINES_IN_SPECTRUM_, to find the best overall ClumpFind
settings that pick up most lines with a range of noise, baseline
wiggles, and residual source emission.  There is also the possibility
of combining the average spectra that exclude detected source emission
from each observation to improve the signal-to-noise and delineate the
reference-spectrum lines within the average spectrum.

It may prove too challenging to fully automate, but the option is
already provided to supply the centres and widths of lines to guide
the removal.