heterodyne: Do not apply source-emission mask and collapse with mode.
The masking can leave residual weak signal, and if the reference
emission lines are embedded within these, it can lead to the reference
line being located in a steep-sided well, which can only partially be
compsensated, and hence it leads to an underestimate of the line's
strength and leaves a residual (albeit much weaker) absorption line in
the reduced spectral cubes.

The mode while COLLAPSE assumes the Pearson relationship for
floating-point data, which does require a well-behaved distribution,
it should at least diminish the source emission (seen here in the
composite spectrum in absorption), as the reference emission wil be
the same in every spectrum bar noise.