heterodyne: Permit analysis using both the masked and unmasked spectra.
This attempts to comnbine the best of both estimates of the reference
spectra.  The masked spectrum is better for identifying the locations
and extents of the lines in the reference spectrum.  The modal
unmasked spectra derives a better estimate of the flux.

To do this make Argument USE_SOURCE_MASK a string rather than boolean,
albeit two of its values remain 1 and 0.  Make the new value of "Both"
the default as this gives the best results thus far.

Also reworded and expanded the description to incorporate this change.