heterodyne: Remove additional assignment for the masked-spectrum background map.
After a new assignment of a work NDF was placed outside of the block
in order that it would be accessible near the end, the original assigment
had not been removed, and so the filled NDF was being erased and the
one expected did not exist.  This wasn't noticed at first because of
another bug affecting the usemask sign (f1682e67cee1a).