heterodyne: Another tweak to the line-fitting settings.
At least one observation (COHRS_15p50_0p00) seem to need these to pick
up the lines.  The worry is that there is no formula that works even
in most cases, and that additional recipe parameters are necessary to
enable more-tailored extraction of the reference spectrum at least in
a few categories.