heterodyne: Process receptor by receptor.
This need for this is obvious as each receptor's vieas a different
reference point will be different.  This may only be small if the
reference point is in a cloud but quite large if there are compact
sources present.  I originally thought that there would not be
sufficient signal, but it seems that was untrue.  Indeed there are quite
large variations between the receptors as evident by the residual grid
pattern in removed referece lines.

Also introduce ITERATION and ONLY_SOURCE_MASK arguments.