cal: Add 600" variant dimmconfig for M16AL003
Renamed the existing (200") dimmconfig to add a "_200" suffix
and adjusted the recpars file accordingly.  This is to make the
difference between the two dimmconfig files clear.  Also added
a new recpars file to select the 600" dimmconfig.  (The existing
recpars file keeps its name as it is in the standard form for
JSA processing.)