JSA: use cloned object to tag as representative
This alters _TAG_AS_REPRESENTATIVE_ so that it works in a temporary
object (frame or group).  This means that it can simply invoke
"force_product_update" rather than having to set the PRODUCT and
PRODID FITS headers manually.  Other headers, such as the bounds
(OBSRATL, OBSRATR, ...) should also now be synced correctly.