general: Restore vector graphic output file
In commit 62267d4c30d776901eaecf1ef1e43c835d495478 it appears that
the variable containing the desired name of the graphic file for vectors
was replaced with the literal string "vector_file".  This caused the
pipeline to store the plot in a file of that name rather than the desired
name.