general: Include DATE headers in graphic header
When a reduction combines data from more than one date (including
a single observation crossing the UT date boundary) the UTDATE
header is removed.  Therefore include the DATE-OBS and DATE-END
in the PNG file header so that we have some record of the date
on which the data were obtained.