SCUBA2: add INVERT option to _THRESHOLD_IMAGE_
The behavior of _THRESHOLD_IMAGE_ seems to normally be to keep the "good"
part of the image unless using the signal-to-noise ratio (COMP=snr)
in which case the "bad" part is kept.  This commit adds a new parameter
"INVERT" which currently applies only when COMP=snr and defaults to
true to retain the current behavior (which appears to be used by
_MASK_SOURCE_EMISSION_).  Calling the primitive with COMP=snr and INVERT=0
should now retain the "good" part of the image based on the
signal-to-noise ratio.