heterodyne: New primitive to rebin spectral cubes.
The reduced cubes are binned to the suppled list of velocity widths,
supplied either through the REBIN argument of recipe parameter.

For most recipes this is done inline within the iterative primitive,
but there are recipes that do not use _ITERATIVE_GROUP_PRODUCTION_ yet
might wish to bin the reduced spectra.  REDUCE_SCIENCE_CONTINUUM is
an example.