cal: Correct incorrect transition names.
Correct a wrong notation of the HARP 338 GHz CH3OH line. In the past
this was changed from 7-6 to the more-correct 7 0 7 +0 - 6 0 6 +0,
which ORAC then changes to 707+0-606+0. This notation was changed when
we added more CH3OH lines to the OT, and index.standard still had the
old names.  This led to reductions not to recognize the transition and
hence integrate over the whole velocity range of the spectrum.