Core: Abort if the list of input files contains a duplicate.
Some recipes can go wrong should it attempt to reduce a file twice.
Although in some cases the origin of the duplication may have been an
extra paste and/or might not cause the recipe to fail, the duplication
may have been an incomplete substitution edit for an additional file.
Therefore it was decided to make file duplication a fatal error rather
than a warning,

This addresses fault 20170509.003.