SCUBA2: Reorder TITLE parameters in SETTITLE commands.
This is to counteract other instances of JCMT Fault 20171123.001
where the ADAM command-line parser is confused by parentheses
in a parameter value, specifically Parameter TITLE setting an NDF's
TITLE component.  It treats them as part of the NDF name, but the
parser operates as expected if the NDF name is supplied last.

Ideally the command-line mangling would be fixed.  Until then
there are likely other instances, whereever a user can set a
string that is passed to an ADAM character variable.