general: New primitive for doubling quotation marks in strings.
This is intended to maniuplate strings that are intended to be passed
to ADAM character-type parameters, such as to assign NDF TITLE and
LABEL components, and which may contain embedded single or double
quotation marks.  This will normally be from a user-supplied value
such as the object name.

The primitive doubles, and escapes if needed, the quotation marks so
that when passed to the ADAM parameter the original string is passed
through the ADAM parameter complete, without truncation after the
quotation mark.

This is a continuation of the resolution of JCMT Fault 20171123.001.
While Commit 143ed721ba053 fixed the obvious crash caused by
parentheses being in a string, after that fix an incomplete object
name was passed to the NDF TITLE compoent.  This routine when applied
should correct that too.