heterodyne: Do not find median Tsys for summit operations.
This addresses Fault 20180108.001, where a summit recipe was taking
more than half of the total processing time to evaluate the median
Tsys.  While having this statistic may prove useful to assess the
quality of an observation, the time taken is too long for other
assessments by the observer.

A further refinement could be to permit calculation of the median Tsys
at the summit when the array has fewer elements than some arbitrary
threshold.