general: Generate _thr file even if object ignored
_THRESHOLD_DATA_ should now generate the _thr file even if the
object name matches the IGNORE parameter, as it already did when
it determined that thresholding was unnecessary.  The reason for
this update is that some group processing recipes rely on the _thr
file being from each frame, but when thresholding was not applied,
the previous file (such as _em) as not kept.  This commit also
has the _thr generated in the case that the calculated median
is bad.