SCUBA2: check if output file list in _FLATFIELD_
Now check that flatfield wrote a (non-empty) output file listing
before attempting to read headers from it.

This primitive is only called from _PROCESS_SCAN_DATA_, which in turn
is only called by REDUCE_SCAN_SUMMIT.  For the summit pipeline, if
delays introduce offsets between the subscan numbers of files found at
the same time, it is possible for it to be called with only fastflat
sequence files.  (Perhaps if, after enough data were already received
for a map to be made, a mixture of science and fastflat files arrive
at once.)  This caused an error where the the pipeline tried to read
FITS headers from undefined files.  For now, trap the problem and issue
a warning but allow the pipeline to continue.