t: Check more instruments in recipe load test
Now find recipes using ORAC::Inst::Defn::orac_determine_recipe_search_path
rather than having to specify a directory for each instrument.  Since
multiple directories can now be included, only the first recipe of each
name found is kept.  For some recipes we need to determine the intended
observing mode in order for the associated primitives to be found
unambiguously.

ORAC_DIR and ORAC_CAL_ROOT are set relative to the current directory even
if they already exist, so that the test checks the current version of the
software.

Several WFCAM and MICHELLE recipes are skipped due to preexisting problems.