heterodyne: Use new _SET_BASELINE_RECPARS_ and add BASELINE_METHOD recpar.
The BASELINE_METHOD is introduced to control where the baseline
regions should come from.  They use a search path of options, the
penultimate being to use recipe parameter BASEDLINE_REGIONS, also
shared with bad-baseline rejection.  This change permits the latter
filtering to use BASELINE_REGIONS while the baseline fitting does not.
Instead it allows the method to be selected.  BASELINE_METHOD
currently only permits the value "auto" (case insensitive) to select
the final option, namely the automated mode where the emission is
detected and masked before baseline fitting.

In regions of complex emission spanning a wide velocity range, using
BASELINE_REGIONS could lead to spectral portions devoid of emission
not being included in the fit, with the potential of undesirable
wiggles in the polynomial fits across the broad gap.