bin: Use xorac_log_window for oracdr_monitor
Now use the ORAC::Xorac::xorac_log_window function to set up the log
panels in the oracdr_monitor application.  This is to avoid the duplicated
code which used to set up the windows here and to take advantage of
recent updates to xorac_log_window.  Support for "results" has also been
added.