bin: Add "--useloc" option to oracdr_monitor
This is to allow oracdr_monitor to use the location file written by
an instance of the pipeline to determine which ORAC_DATA_OUT directory
to use.  The check for ORAC_DATA_OUT being defined now has to take place
after the command line options are processed in case --useloc is specified.