lib: Restore $$ to KAPVIEW device name
This is to allow multiple pipelines to be run in the same session.
(E.g. running the SCUBA-2 850um and 450um pipelines together.)
However the device name suffix can now be specified by a new environment
variable ORAC_KAPVIEW_DEV in case it is desired that the pipeline reuse
an existing set of KAPVIEW windows. (Note $$ was previously included
in the device name until it was removed in commit f31f6105a6 but it is
unknown why that was done.)