bin: Avoid using Tk::fileevent in oracdr_monitor
This does not seem to work well with actual files, triggering all the
time, so that the script relied on the timing of it rescheduling the
fileevents when it read nothing.  It might as well just read the file
in a timer event.  This solves the problem of the KAPVIEW display
flickering and ending up incomplete when several files are displayed
in quick succession, such as SCUBA-2 focus images.