bin: Only seek to the end of initial display file
The oracdr_monitor script skips to the end of new display files so that
it doesn't receive a large number of display requests when it starts up.
Now it should only do this the first time -- for new display files which it
detects, it starts from the beginning so as not to miss the first display
request in the file.