bin: Override GAIA history dir. for oracdr_monitor
We should not use ORAC_DATA_OUT for this in the case of oracdr_monitor
in order to avoid interfering with the active pipeline.  This could
perhaps have used an existing variable (such as ORACDR_TMP or maybe
even RTD_REMOTE_DIR) but using a new variable seemed safer in case of
unexpected usage of those other variables.