hetrodyne: Add ability to use a reference file to define the output grid.
This adds a new REFERENCE argument and a corresponding recipe parameter.
A valid reference file is supplied to MAKECUBE to set the WCS of the
output spectral cube.

It involved some code re-organising.  Hence looking at some of the
individual difference chunks may be less helpful than viewing the
code.