heterodyne: Restructure code for the reference file code.
Improve the structure and make the ordering more logical in the light
of the reference-file facility.  Rather than having lots of tests for
"! $useref" in normal processing, try to block together related code
that only applies in the normal, non-reference-file, state.  The main
related code being the WCS grid parameters.  Some of these WCS
parameters were being calculated then thrown away in the initial
implementation of the reference-file functionality.