heterodyne: Invoke the new line-forest.
This may be controversial in that disables the FOREST flag in
MFITTREND if the observation turns out not to be of a line forest, and
a warning that this has been done.  This is to prevent biased or
spurious baseline subtractions because the aggressive forest filtering
introduced in 2016 masks too much spectrum in ordinary spectra.

While I'm normally against software undermining what the user
selected, here I think it has some merit, although my main goal was to
prevent nonsense baseline subtractions in the JSA processing.  In the
early promotion of ORAC-DR, it was described as a "data-driven"
pipeline.  This seems to a case where the data drives the appropriate
processing...  Just to note that the rest of the
REDUCE_SCIENCE_LINEFOREST is unchanged.