heterodyne: Issue an expanded warning leaving LINEFOREST_BASELINE unchanged.
After discussions on Scicom, it was felt that switching the FOREST
parameter in MFITTREND was too magical for users, albeit data-driven,
and we should use a different approach to correct the matter of
incorrect recipe selection leading to bad baseline subtraction in the
JSA processing.

Expand the warning to users to include the recipe parameter, and
briefly why the current setting or eceipe is inappropriate.