audio: Add warnings for the QL having stopped
This adds two audio files to be used as warnings when the SCUBA-2
quick-look pipelines stop due to an error.  This is done by including
a new "audiosuffix" option in the oracdr alias.  Also the environment
variable PULSE_SERVER is set if the REMOTEHOST is one of the dedicated
summit pipeline display machines.