heterodyne: Permit alignment between sidebands.
This allows data taken in different sidebands to be aligned in
MAKECUBE.  An example ocurs with COHRS. The 2018 observations were
observed in the lower sideband, whereas the original COHRS survey was
observed in the upper sideband.  In order to combine old and new data
in ORAC-DR to form a tile of higher signal to noise, the AST attribute
AlignSideBand needed to be set to 1.