_CALCULATE_STANDARD_: add recpar and primitive options
This is so that in the normal recipes _CALCULATE_STANDARD_ won't run
by default if the source is not known to the calibration
system. However, the user can choose to force it to run on everything
by using the CALCULATE_STANDARD_ALWAYS=1 recpar option.

The picard recipe CALCULATE_STANDARD and REDUCE_STANDARD use the
primitive option to ensure that _CALCULATE_STANDARD_ will always run,
even if the source isn't known.