CALCULATE_STANDARD should run on 1,1 and error if there is no data there
Previously it would use FIND_TRUE_PEAK and default to the peak pixel
if there was no data in the 1,1 position. However, this is really only
designed for stares, and generally the standard value for an adjacent
receptor is not helpful. This may need to be revisited after the JCMT
calibration database re runs the historical acsis data, if it turns
out 1,1 isn't always the location of the source.