recpars: JCMTCAL observations use looser baseline linearity clipping
I suspect this should be the default for HARP stares/spare grids with
small numbers of positions -- if each receptor is only looking at one
posiiton, then the linearity clipping needs to be very broad, as often
a source is only in one or a small number of positions/detectors. The
problem was particulary bad for the JCMTCAL spectral standards with
HARP, where often one receptor has a huge bright line and all the
others have nothing. The baseline linearity clipping was frequently
masking out the source pixel.