PICARD: add a recipe to create a HiPS structure
This recipe was originally developed for the JCMT Transient large program
and is being added to the pipeline in case it is useful elsewhere.

The _MAKE_HIPS_ primitive is extremely slow due to the amount of I/O
performed to do this using KAPPA tasks.  It is also currently limited
to a single resolution.  This is because jsadicer can only be controlled
via an "instrument" parameter which chooses one of a number of fixed
tiling schemes and because the primitive assumes it will be given
1024x1024 tiles, meaning that only one of those schemes is suitable.
Ideally it would be possible to specify any desired HiPS order.