lib: Compare date recpar filter headers as ISO
If a recpar filter is applied to a header value which is a datetime
(i.e. it has a "datetime" method), it is converted to ISO format for
comparison (using that "datetime" method).  This is to allow such
headers to be written in a manner similar to that in the raw FITS headers.