heterodyne: Add new primitive for post-processing reference-spectrum removal.
This works removing most of the unwanted signal, but it is not the
optimum way, as noted in its Description, as residual absorption lines
can affect the baseline fitting and flat-fielding.  In the worst cases
where I have applied this, there is the familiar grid pattern at the
velocities where a reference line was located.  If a more-rigorous
technique fails to deliver, say because of noise affecting the
interpolation across the locations of reference-spectrum lines, I may
come back to this.  Committing this primitive for security.