heterodyne: Create a new simplified removal of reference-spectrum lines.
This new primitive interpolates across reference absorption lines
using KAPPA:FILLBAD that are located by a user-specified list of
velocity extents.  It can derive and subtract the reference spectra by
receptor or combine them if the noise is too high for the former.
It seems to give acceptable results leading to better flat fielding
across the line extents, instead of a residual grid pattern.