heterodyne: Improve the file tidying after removing reference emission.
When operating the reference-emission removal code, lots of
intermediate files remain.  That's in part because the method is only
partially successful, and these residual files were useful for
debugging, (although it's still not clear why 11% of lines are
unchanged, and 57% are only partially filled).  Since I want to
re-reduce many such observations with an extra correction primitive
that does remove the residual absorption features, it's time to clear
up the mess. The handful of files per receptor can always be restored
with ORAC_KEEP enabled.

The main steps were calling the appropriate Group or Frame
push_intermediates{} method, including in new frame contexts; and not
passing TempFile files as a new Frame's file to subroutine-like
primitives that invoke inout(), but instead make regular Frame
filenames.  However, the _resmask file stubbornly refuses to go, so a
bug remains, albeit leaving a relatively small file.