heterodyne: Improve the file tidying after removing reference emission.
This is a continuation of Commit 55b474ea633a9.  The residual
intermediate file created in _MASK_SINGLE_AXIS_REGIONS_ was extremely
reluctant to be removed, hence could not be part of the original
commit.  Upon returning to the matter three weeks later a solution
was found, albeit it was an educated guess rather than logic.