heterodyne: Use consistent parameter names.
Since nobody has objected to this proposed change after ten days...
See Fault 20200130.009.

For some unknown reason (dated back to 2009 October or earlier) the
recipe parameter names for the velocity limits applied to computing
moments were different at the user level from those accessed by the
primitive.  This is abnormal.  It led to confusion for a user because
the PICARD recipe CREATE_MOMENTS_MAP that invokes
_CREATE_MOMENTS_MAPS_THROUGH_SMOOTHING_ adapted its documentation from
this primitive.

I had been aware of this since c.2012 and wanted to make this
change long ago, but feared that it might break something.  Of course,
the original naming might have been a mistake rather than intentional.