Core: Log the Starlink version and STARLINK_DIR.
Many times when examining an ORAC-DR logfile I've wanted to know which
version of Starlink provided the engines for ORAC-DR.  It can be as
important for understanding different behaviour in a series of
reductions as the ORAC-DR version.

Not including Starlink in the log preamble has longed seemed an
anomaly to me.  We can revert this commit if it violates some ORAC-DR
principle.  Yes the underlying Starlink engines could be replaced, but
it's highly unlikely now, especially given EAO's commitment to
continue using Starlink.  If they were replaced, it's a minor edit to
add information about the new applications code.