heterodyne: create log.receptor
This is a log file with average RMS and TSYS per receptor for each
frame. There have been some requests from instrument
scientists/engineers to log this in the past, and as it is already
calculated for the log.qa, I have created a machine-readable
log.receptor file for it. I'm not sure how useful it will be to the
engineers however, as this is calculated after data reduction has removed
problems.