heterodyne: Add Quiet argumen to _CALCULATE_STANDARD_.
When _CALCULATE_STANDARD_ is called from recipes that process
non-standard sources, notably the change at Commit 709a7f63b55a1, it
will always generate a warning that the source is not a known
standard.  This can cause unnecessary consternation to users.  While
processing in REDUCE_STANDARD such warnings are pertinent.  To overcome
this dichotomy, this commit adds a new argmument called QUIET, which, if
set true, prevents the warning and additional commentary.