heterodyne: Invoke new QUIET argument of _CALCULATE_STANDARD_.
This prevents an annoying and possibly troubling warning when the
source is not a standard, as will normally be the case when processed
by _ITERATIVE_GROUP_PRODUCTION_.  If however, a standard is reduced
with a recipe other than REDUCE_STANDARD, the modified code will not
be reached.  If a source is not a known standard is processed with
REDUCE_STANDARD, the warning will still be generated.