heterodyne: Create master reference spectra, aligning pixels.
A small number of COHRS reductions were not removing the residual
reference signal, and indeed one added some emission.  This was due
to combining observations taken on different dates, and not having the
same mapping from VRAD to pixels.  Hence the reference spectrum was
blurred at best or underestimated.  These modificatiins align those
spectra before to make a master reference spectrum for each group.

Also some artefacts arose because the reference position was not
exactly the same for data taken on different nights.  This could
appear as small shifts in the velocity of the reference lines, giving
artefactul ripples in the master reference spectrum.

When subtracting the master reference spectrum for each reference
position or UT date, alignment with each observation is also undertaken
so then a simple subtraction can be used to filter the reference signal
from that observation.

The grpFrm is replaced by the clearer  masterFrm Frame.