heterodyne: First attempt at alignment in _SUBTRACT_REFERENCE_EMISSION_.
This, in esssence, applies similar changes to those made to
_SUBTRACT_REFERENCE_SPECTRUM_ in order to account for different velocity
origins, aligning to the same spectral scale.  This introduces a
masterFrm Frame

It is only partially successful, in that the alignment to form a
master reference spectrum works, but the realignment of the esxpanded
reference spectrum as a time-series cube to the input cube is suspect.
I wanted to save the development before a long weekend.