heterodyne: Propagate the original WCS to the expanded reference spectra.
Without this step the aligned time-series cube form of the reference spectra
are mostly bad.  It would be if the master reference emission approach
could be made to work, say with WCSALIGN ALIGNREF=F and an AST
attribute set appropriately...

I had tried restoring the third-axis (time) attributes with WCSATTRIB
MODE=MSET, but the whole Frame was needed.