heterodyne: Re-order alignment in application phase of off-signal removal
In the application phase, instead of trying to pixel align a reference
spectrum grown to the original current cube's dimensions, extract a
spectrum from the current data file and align the reference spectrum
to that, then copy that to all times and receptors in the original
time-series cube.  The original approach gave insurmountable problems
propagating the WCS.  The new approach is also faster, only aligning
one spectrum rather than potentially tens of thousands of spectra.

Adjust the file tidying for the new filenames and for old ones no
longer created.