heterodyne: Augment and indent the log commentary for ref. spectrum removal
Improve the log reporting so that all the sub-files are listed as they
are processed, and related operations under an introductory comment
are indented to show the steps more clearly.  Also add in some line
feeds to separate sections.  There were operations being reported on
files not previously mentioned, but now there are log messages saying
what these are.

Indent the log message in _ALIGN_SPECTRUM too to make the above work.
Although it does assume some connection with the calling primitives,
it will be normal to have some general log message and then call
_ALIGN_SPECTRA_, perhaps on several files, so indenting in
_ALIGN_SPECTRA_ seems reasonable.  For generic primitives like this,
it might be better to offer a log-commentary indentation argument
specifiying the number of characters to indent.

Also fix punctuation in the same comment in both _SUBTRACT_REFERENCE_...
files.